ocr_sandbox/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
pdf2image
python-multipart
